RenAIssance_Transformer_OCR_Utsav_Rai/requirements.txt
# Basic Python packages
numpy==1.26.4
Pillow==10.2.0
opencv-python==4.9.0.80
scikit-image==0.22.0
deskew==1.5.1
fitz==0.0.1.dev2
PyMuPDF==1.24.4
argparse==1.4.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.39.1
datasets==2.18.0
matplotlib==3.8.3
scikit-learn==1.4.2
pyyaml==6.0.1
wandb==0.17.1
python-Levenshtein==0.25.0
scipy==1.13.0